Move discovery into an internal module
It's still pulled in by API components, and still exposed in some
interfaces, but now largely internalized by re-namespacing.
