+++ src/mddj/api/reader/_config.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import dataclasses
 import pathlib
 
-from ..discovery import DirExplorer
+from ..._internal import _discovery
 
 
 @dataclasses.dataclass
@@ -12,7 +12,7 @@ class ReaderConfig:
     Configuration for a metadata reader.
     """
 
-    dir_explorer: DirExplorer
+    dir_explorer: _discovery.DirExplorer
     project_directory: pathlib.Path | None
     isolated_builds: bool
     capture_build_output: bool
